+++ packages/kbn-reporting/export_types/pdf/printable_pdf_v2.test.ts
@@ -1,24 +1,27 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
+
+import { of } from 'rxjs';
+import type { Writable } from 'stream';
 
 import { coreMock, elasticsearchServiceMock, loggingSystemMock } from '@kbn/core/server/mocks';
 import { CancellationToken } from '@kbn/reporting-common';
-import { PdfV1ExportType } from '@kbn/reporting-export-types-pdf';
-import { TaskPayloadPDF } from '@kbn/reporting-export-types-pdf-common';
-import { cryptoFactory, generatePdfObservable } from '@kbn/reporting-server';
-import { ScreenshottingStart } from '@kbn/screenshotting-plugin/server';
-import { of } from 'rxjs';
-import { Writable } from 'stream';
-import { createMockConfigSchema } from '../../test_helpers';
+import type { LocatorParams } from '@kbn/reporting-common/types';
+import type { TaskPayloadPDFV2 } from '@kbn/reporting-export-types-pdf-common';
+import { cryptoFactory, generatePdfObservableV2 } from '@kbn/reporting-server';
+import type { ScreenshottingStart } from '@kbn/screenshotting-plugin/server';
 
-jest.mock('@kbn/reporting-server/generate_pdf');
+import { PdfExportType } from '.';
+
+jest.mock('@kbn/reporting-server/generate_pdf_v2');
 
 let content: string;
-let mockPdfExportType: PdfV1ExportType;
+let mockPdfExportType: PdfExportType;
 let stream: jest.Mocked<Writable>;
 
 const cancellationToken = {
@@ -33,18 +36,22 @@ const encryptHeaders = async (headers: Record<string, string>) => {
   return await crypto.encrypt(headers);
 };
 
-const getBasePayload = (baseObj: any) => baseObj as TaskPayloadPDF;
+const getBasePayload = (baseObj: any) =>
+  ({
+    params: { forceNow: 'test' },
+    ...baseObj,
+  } as TaskPayloadPDFV2);
 
 beforeEach(async () => {
   content = '';
   stream = { write: jest.fn((chunk) => (content += chunk)) } as unknown as typeof stream;
+
   const configType = createMockConfigSchema({ encryptionKey: mockEncryptionKey });
   const context = coreMock.createPluginInitializerContext(configType);
 
   const mockCoreSetup = coreMock.createSetup();
   const mockCoreStart = coreMock.createStart();
-
-  mockPdfExportType = new PdfV1ExportType(mockCoreSetup, configType, mockLogger, context);
+  mockPdfExportType = new PdfExportType(mockCoreSetup, configType, mockLogger, context);
 
   mockPdfExportType.setup({
     basePath: { set: jest.fn() },
@@ -57,25 +64,31 @@ beforeEach(async () => {
   });
 });
 
-afterEach(() => (generatePdfObservable as jest.Mock).mockReset());
+afterEach(() => (generatePdfObservableV2 as jest.Mock).mockReset());
 
 test(`passes browserTimezone to generatePdf`, async () => {
   const encryptedHeaders = await encryptHeaders({});
-  (generatePdfObservable as jest.Mock).mockReturnValue(of({ buffer: Buffer.from('') }));
+  (generatePdfObservableV2 as jest.Mock).mockReturnValue(of(Buffer.from('')));
 
   const browserTimezone = 'UTC';
   await mockPdfExportType.runTask(
     'pdfJobId',
     getBasePayload({
+      forceNow: 'test',
+      title: 'PDF Params Timezone Test',
+      locatorParams: [{ version: 'test', id: 'test' }] as LocatorParams[],
       browserTimezone,
       headers: encryptedHeaders,
-      objects: [{ relativeUrl: '/app/kibana#/something' }],
     }),
     cancellationToken,
     stream
   );
 
-  expect(generatePdfObservable).toHaveBeenCalledWith(
+  expect(generatePdfObservableV2).toHaveBeenCalledWith(
+    expect.anything(),
+    expect.anything(),
+    expect.anything(),
+    expect.anything(),
     expect.anything(),
     expect.objectContaining({ browserTimezone: 'UTC' })
   );
@@ -84,11 +97,11 @@ test(`passes browserTimezone to generatePdf`, async () => {
 test(`returns content_type of application/pdf`, async () => {
   const encryptedHeaders = await encryptHeaders({});
 
-  (generatePdfObservable as jest.Mock).mockReturnValue(of({ buffer: Buffer.from('') }));
+  (generatePdfObservableV2 as jest.Mock).mockReturnValue(of({ buffer: Buffer.from('') }));
 
   const { content_type: contentType } = await mockPdfExportType.runTask(
     'pdfJobId',
-    getBasePayload({ objects: [], headers: encryptedHeaders }),
+    getBasePayload({ locatorParams: [], headers: encryptedHeaders }),
     cancellationToken,
     stream
   );
@@ -97,12 +110,12 @@ test(`returns content_type of application/pdf`, async () => {
 
 test(`returns content of generatePdf getBuffer base64 encoded`, async () => {
   const testContent = 'test content';
-  (generatePdfObservable as jest.Mock).mockReturnValue(of({ buffer: Buffer.from(testContent) }));
+  (generatePdfObservableV2 as jest.Mock).mockReturnValue(of({ buffer: Buffer.from(testContent) }));
 
   const encryptedHeaders = await encryptHeaders({});
   await mockPdfExportType.runTask(
     'pdfJobId',
-    getBasePayload({ objects: [], headers: encryptedHeaders }),
+    getBasePayload({ locatorParams: [], headers: encryptedHeaders }),
     cancellationToken,
     stream
   );
