+++ packages/kbn-reporting/export_types/png/png_v2.test.ts
@@ -1,8 +1,9 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
 import * as Rx from 'rxjs';
@@ -11,12 +12,11 @@ import { Writable } from 'stream';
 import { coreMock, elasticsearchServiceMock, loggingSystemMock } from '@kbn/core/server/mocks';
 import { CancellationToken } from '@kbn/reporting-common';
 import type { LocatorParams } from '@kbn/reporting-common/types';
-import { PngExportType } from '@kbn/reporting-export-types-png';
 import type { TaskPayloadPNGV2 } from '@kbn/reporting-export-types-png-common';
 import { cryptoFactory, generatePngObservable } from '@kbn/reporting-server';
 import type { ScreenshottingStart } from '@kbn/screenshotting-plugin/server';
 
-import { createMockConfigSchema } from '../../test_helpers';
+import { PngExportType } from '.';
 
 jest.mock('@kbn/reporting-server/generate_png');
